60FPS + more time to drop
Force rendering at 60FPS max: this increases performance dramastically!
At least a factor 2, sometimes 4-5...
Everything comes with this variable get_rs()->min_framerate which gives
a value of 1000... I tried a lot of different things, but if I change
this value to 60 then the game lags. My best fix is the one I put in
this commit. I will continue to investigate but this is already a very
good step!

Also, I multiplied by 5 the time to drop an item, to avoid frustration
:)