New skin for the 2nd toolbar
This commit adds:
- better info when you look at the skills in your quick toolbar
- fix some non important stuff in the code to make it cleaner
- ask for an other xml file when one launches the 2nd skill bar.

To change the skins:
- there is an xml file read by the code to tell which file it should
load and at wich position. It also gives number to be able to call a
specific part of the "skin" in the code.
- then this xml files reference a name and there is a link between this
name and a file (like UI_strID.id). In this file you have to write the
name and the line in the corresponding  .TSI file. BE SUPER CAREFUL when
editing the .id file. The space are weird and super important! (it has
to be a \r\n between each line!!!)
- then you just have to edit the TSI file to add the dds