New basic skill - gather at distance and correct use of mana/stamina
Addition of a new basic skill: gather at distance for a cost of 1000
stamina.
Furthermore, I added the correct consumption of mana/energy/requirement
(wing equipped :)) to basic skill. For the consumption, I had to add it
both in client and in the packet on the server side.
Now floating jump skill needs wings and 4mana.