Double chat and more
This commit adds a lot of functionnality in the code to play with chat
objects and XML.
In particular, I've updated it so it now allows various size for the
chat and separetes the the system chat, clan chat, whisper chat etc
(thanks to the filter or tabs). For this one needs to take the new UI
and XML files.
Furthermore, there are new commands in the chat (like the /sort one to
sort your inventory!). To learn about them just write /help in the chat.